Inline release steps in release command base class
